refactor(hlo): move backend code closer to models
