+++ optimum/neuron/models/inference/hlo/llama/modules.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-from .....backends.hlo import module
+from ..backend import module
 
 
 class LlamaForCausalLM(module.PretrainedModel):
