+++ optimum/neuron/models/inference/hlo/qwen2/model.py
@@ -15,8 +15,8 @@
 
 from transformers import PretrainedConfig
 
-from .....backends.hlo.config import HloNeuronConfig
-from .....backends.hlo.dtypes import to_torch_dtype
+from ..backend.config import HloNeuronConfig
+from ..backend.dtypes import to_torch_dtype
 from ..llama.model import LlamaHloModel
 from .modules import Qwen2ForCausalLM
